dom/xbl/test/test_bug481558.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=481558
-->
<head>
  <title>Test for Bug 481558</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
  <link rel="stylesheet" type="text/css"
        href="file_bug481558css.sjs?id=id1">
  <link rel="stylesheet" type="text/css"
        href="file_bug481558css.sjs?id=id2&server=example.com">
  <link rel="stylesheet" type="text/css" href="http://example.com/tests/dom/xbl/test/file_bug481558css.sjs?id=id3">
  <link rel="stylesheet" type="text/css" href="http://example.com/tests/dom/xbl/test/file_bug481558css.sjs?id=id4&server=localhost:8888">
</head>
<body onload="runTest();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=481558">Mozilla Bug 481558</a>
<p id="id1"></p>
<p id="id2"></p>
<p id="id3"></p>
<p id="id4"></p>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
function runTest() {
  is ($('id1').xblBoundProperty, 1, "XBL should be attached");
  is ($('id2').xblBoundProperty, undefined, "XBL shouldn't be attached");
  is ($('id3').xblBoundProperty, undefined, "XBL shouldn't be attached");
  is ($('id4').xblBoundProperty, undefined, "XBL shouldn't be attached");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
